+++ app/controllers/projects_controller.rb
@@ -268,7 +268,7 @@ class ProjectsController < ApplicationController
       :show_contacts, :show_documents, :show_posts, :disable_all_emails,
       :collect_email_on_surveys, :lockable, :hide_values_on_pdfs,
       :double_data_entry, :randomizations_enabled, :adverse_events_enabled,
-      :blinding_enabled,
+      :blinding_enabled, :handoffs_enabled,
       # Uploaded Logo
       :logo, :logo_uploaded_at, :logo_cache, :remove_logo,
       # Will automatically generate a site if the project has no site