+++ app/models/project.rb
@@ -51,6 +51,7 @@ class Project < ActiveRecord::Base
 
   has_many :exports, -> { where deleted: false }
   has_many :events, -> { where deleted: false }
+  has_many :handoffs
 
   has_many :contacts, -> { where deleted: false }
   has_many :documents, -> { where deleted: false }