+++ db/migrate/20151109183545_add_handoffs_enabled_to_projects.rb
@@ -1,0 +1,5 @@
+class AddHandoffsEnabledToProjects < ActiveRecord::Migration
+  def change
+    add_column :projects, :handoffs_enabled, :boolean, null: false, default: false
+  end
+end