+++ test/controllers/handoff_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class HandoffControllerTest < ActionController::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end